+++ src/cbor/common.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014-2015 Pavel Kalvoda <me@pavelkalvoda.com>
+ * Copyright (c) 2014-2019 Pavel Kalvoda <me@pavelkalvoda.com>
  *
  * libcbor is free software; you can redistribute it and/or modify
  * it under the terms of the MIT license. See LICENSE for details.
@@ -19,7 +19,8 @@
 #ifdef __cplusplus
 extern "C" {
 
-/** C++ is not a subset of C99 -- 'restrict' qualifier is not a part of the language.
+/**
+ * C++ is not a subset of C99 -- 'restrict' qualifier is not a part of the language.
  * This is a workaround to keep it in C headers -- compilers allow linking non-restrict
  * signatures with restrict implementations.
  *
@@ -55,6 +56,21 @@ static const uint8_t cbor_patch_version = CBOR_PATCH_VERSION;
 
 #define TO_STR_(x) #x
 #define TO_STR(x) TO_STR_(x) /* enables proper double expansion */
+
+// Macro to short-circuit builder functions when memory allocation fails
+#define _CBOR_NOTNULL(cbor_item) do { \
+		if (cbor_item == NULL) { \
+			return NULL; \
+		} \
+	} while (0)
+
+// Macro to short-circuit builders when memory allocation of nested data fails
+#define _CBOR_DEPENDENT_NOTNULL(cbor_item, pointer) do { \
+		if (pointer == NULL) { \
+			_CBOR_FREE(cbor_item); \
+			return NULL; \
+		} \
+	} while (0)
 
 #if CBOR_CUSTOM_ALLOC