+++ src/cbor/floats_ctrls.c
@@ -49,14 +49,18 @@ double cbor_float_get_float8(const cbor_item_t *item)
 	return *(double *) item->data;
 }
 
-double cbor_float_get_float(const cbor_item_t * item)
+double cbor_float_get_float(const cbor_item_t *item)
 {
 	assert(cbor_is_float(item));
-	switch(cbor_float_get_width(item)) {
-	case CBOR_FLOAT_0: return NAN;
-	case CBOR_FLOAT_16: return cbor_float_get_float2(item);
-	case CBOR_FLOAT_32: return cbor_float_get_float4(item);
-	case CBOR_FLOAT_64: return cbor_float_get_float8(item);
+	switch (cbor_float_get_width(item)) {
+		case CBOR_FLOAT_0:
+			return NAN;
+		case CBOR_FLOAT_16:
+			return cbor_float_get_float2(item);
+		case CBOR_FLOAT_32:
+			return cbor_float_get_float4(item);
+		case CBOR_FLOAT_64:
+			return cbor_float_get_float8(item);
 	}
 	return NAN; /* Compiler complaints */
 }
@@ -98,11 +102,13 @@ bool cbor_ctrl_is_bool(const cbor_item_t *item)
 cbor_item_t *cbor_new_ctrl()
 {
 	cbor_item_t *item = _CBOR_MALLOC(sizeof(cbor_item_t));
+	_CBOR_NOTNULL(item);
+
 	*item = (cbor_item_t) {
-		.type = CBOR_TYPE_FLOAT_CTRL,
-		.data = NULL,
-		.refcount = 1,
-		.metadata = {.float_ctrl_metadata = {.width = CBOR_FLOAT_0, .ctrl = CBOR_CTRL_NONE}}
+			.type = CBOR_TYPE_FLOAT_CTRL,
+			.data = NULL,
+			.refcount = 1,
+			.metadata = {.float_ctrl_metadata = {.width = CBOR_FLOAT_0, .ctrl = CBOR_CTRL_NONE}}
 	};
 	return item;
 }
@@ -110,11 +116,13 @@ cbor_item_t *cbor_new_ctrl()
 cbor_item_t *cbor_new_float2()
 {
 	cbor_item_t *item = _CBOR_MALLOC(sizeof(cbor_item_t) + 4);
+	_CBOR_NOTNULL(item);
+
 	*item = (cbor_item_t) {
-		.type = CBOR_TYPE_FLOAT_CTRL,
-		.data = (unsigned char *) item + sizeof(cbor_item_t),
-		.refcount = 1,
-		.metadata = {.float_ctrl_metadata = {.width = CBOR_FLOAT_16}}
+			.type = CBOR_TYPE_FLOAT_CTRL,
+			.data = (unsigned char *) item + sizeof(cbor_item_t),
+			.refcount = 1,
+			.metadata = {.float_ctrl_metadata = {.width = CBOR_FLOAT_16}}
 	};
 	return item;
 }
@@ -122,11 +130,13 @@ cbor_item_t *cbor_new_float2()
 cbor_item_t *cbor_new_float4()
 {
 	cbor_item_t *item = _CBOR_MALLOC(sizeof(cbor_item_t) + 4);
+	_CBOR_NOTNULL(item);
+
 	*item = (cbor_item_t) {
-		.type = CBOR_TYPE_FLOAT_CTRL,
-		.data = (unsigned char *) item + sizeof(cbor_item_t),
-		.refcount = 1,
-		.metadata = {.float_ctrl_metadata = {.width = CBOR_FLOAT_32}}
+			.type = CBOR_TYPE_FLOAT_CTRL,
+			.data = (unsigned char *) item + sizeof(cbor_item_t),
+			.refcount = 1,
+			.metadata = {.float_ctrl_metadata = {.width = CBOR_FLOAT_32}}
 	};
 	return item;
 }
@@ -134,11 +144,13 @@ cbor_item_t *cbor_new_float4()
 cbor_item_t *cbor_new_float8()
 {
 	cbor_item_t *item = _CBOR_MALLOC(sizeof(cbor_item_t) + 8);
+	_CBOR_NOTNULL(item);
+
 	*item = (cbor_item_t) {
-		.type = CBOR_TYPE_FLOAT_CTRL,
-		.data = (unsigned char *) item + sizeof(cbor_item_t),
-		.refcount = 1,
-		.metadata = {.float_ctrl_metadata = {.width = CBOR_FLOAT_64}}
+			.type = CBOR_TYPE_FLOAT_CTRL,
+			.data = (unsigned char *) item + sizeof(cbor_item_t),
+			.refcount = 1,
+			.metadata = {.float_ctrl_metadata = {.width = CBOR_FLOAT_64}}
 	};
 	return item;
 }
@@ -146,6 +158,7 @@ cbor_item_t *cbor_new_float8()
 cbor_item_t *cbor_new_null()
 {
 	cbor_item_t *item = cbor_new_ctrl();
+	_CBOR_NOTNULL(item);
 	cbor_set_ctrl(item, CBOR_CTRL_NULL);
 	return item;
 }
@@ -153,6 +166,7 @@ cbor_item_t *cbor_new_null()
 cbor_item_t *cbor_new_undef()
 {
 	cbor_item_t *item = cbor_new_ctrl();
+	_CBOR_NOTNULL(item);
 	cbor_set_ctrl(item, CBOR_CTRL_UNDEF);
 	return item;
 }
@@ -165,6 +179,7 @@ cbor_item_t *cbor_build_bool(bool value)
 cbor_item_t *cbor_build_float2(float value)
 {
 	cbor_item_t *item = cbor_new_float2();
+	_CBOR_NOTNULL(item);
 	cbor_set_float2(item, value);
 	return item;
 }
@@ -172,6 +187,7 @@ cbor_item_t *cbor_build_float2(float value)
 cbor_item_t *cbor_build_float4(float value)
 {
 	cbor_item_t *item = cbor_new_float4();
+	_CBOR_NOTNULL(item);
 	cbor_set_float4(item, value);
 	return item;
 }
@@ -179,13 +195,15 @@ cbor_item_t *cbor_build_float4(float value)
 cbor_item_t *cbor_build_float8(double value)
 {
 	cbor_item_t *item = cbor_new_float8();
+	_CBOR_NOTNULL(item);
 	cbor_set_float8(item, value);
 	return item;
 }
 
-cbor_item_t *cbor_build_ctrl(uint8_t  value)
+cbor_item_t *cbor_build_ctrl(uint8_t value)
 {
 	cbor_item_t *item = cbor_new_ctrl();
+	_CBOR_NOTNULL(item);
 	cbor_set_ctrl(item, value);
 	return item;
 }