+++ src/cbor/floats_ctrls.h
@@ -74,7 +74,7 @@ double cbor_float_get_float(const cbor_item_t * item);
  *
  * The width cannot be changed once the item is created
  *
- * @return **new** 1B ctrl
+ * @return **new** 1B ctrl or `NULL` upon memory allocation failure
  */
 cbor_item_t * cbor_new_ctrl();
 
@@ -82,7 +82,7 @@ cbor_item_t * cbor_new_ctrl();
  *
  * The width cannot be changed once the item is created
  *
- * @return **new** 2B float
+ * @return **new** 2B float or `NULL` upon memory allocation failure
  */
 cbor_item_t * cbor_new_float2();
 
@@ -90,7 +90,7 @@ cbor_item_t * cbor_new_float2();
  *
  * The width cannot be changed once the item is created
  *
- * @return **new** 4B float
+ * @return **new** 4B float or `NULL` upon memory allocation failure
  */
 cbor_item_t * cbor_new_float4();
 
@@ -98,37 +98,39 @@ cbor_item_t * cbor_new_float4();
  *
  * The width cannot be changed once the item is created
  *
- * @return **new** 8B float
+ * @return **new** 8B float or `NULL` upon memory allocation failure
  */
 cbor_item_t * cbor_new_float8();
 
 /** Constructs new null ctrl item
  *
- * @return **new** null ctrl item
+ * @return **new** null ctrl item or `NULL` upon memory allocation failure
  */
 cbor_item_t * cbor_new_null();
 
 /** Constructs new undef ctrl item
  *
- * @return **new** undef ctrl item
+ * @return **new** undef ctrl item or `NULL` upon memory allocation failure
  */
 cbor_item_t * cbor_new_undef();
 
 /** Constructs new boolean ctrl item
  *
  * @param value The value to use
- * @return **new** boolen ctrl item
+ * @return **new** boolen ctrl item or `NULL` upon memory allocation failure
  */
 cbor_item_t * cbor_build_bool(bool value);
 
 /** Assign a control value
  *
  * \rst
- * .. warning:: It is possible to produce an invalid CBOR value by assigning a invalid value using this mechanism. Please consult the standard before use.
+ * .. warning:: It is possible to produce an invalid CBOR value by assigning a
+ * invalid value using this mechanism. Please consult the standard before use.
  * \endrst
  *
  * @param item[borrow] A ctrl item
- * @param value The simple value to assign. Please consult the standard for allowed values
+ * @param value The simple value to assign. Please consult the standard for
+ * 	allowed values
  */
 void cbor_set_ctrl(cbor_item_t * item, uint8_t value);
 
@@ -177,21 +179,21 @@ cbor_item_t *cbor_build_float2(float value);
 /** Constructs a new float
  *
  * @param value the value to use
- * @return **new** float
+ * @return **new** float or `NULL` upon memory allocation failure
  */
 cbor_item_t *cbor_build_float4(float value);
 
 /** Constructs a new float
  *
  * @param value the value to use
- * @return **new** float
+ * @return **new** float or `NULL` upon memory allocation failure
  */
 cbor_item_t *cbor_build_float8(double value);
 
 /** Constructs a ctrl item
  *
  * @param value the value to use
- * @return **new** ctrl item
+ * @return **new** ctrl item or `NULL` upon memory allocation failure
  */
 cbor_item_t *cbor_build_ctrl(uint8_t  value);