+++ src/cbor/maps.c
@@ -23,9 +23,8 @@ size_t cbor_map_allocated(const cbor_item_t *item)
 cbor_item_t *cbor_new_definite_map(size_t size)
 {
 	cbor_item_t *item = _CBOR_MALLOC(sizeof(cbor_item_t));
-	if (item == NULL) {
-		return NULL;
-	}
+	_CBOR_NOTNULL(item);
+
 	*item = (cbor_item_t) {
 		.refcount = 1,
 		.type = CBOR_TYPE_MAP,
@@ -36,18 +35,16 @@ cbor_item_t *cbor_new_definite_map(size_t size)
 		}},
 		.data = _cbor_alloc_multiple(sizeof(struct cbor_pair), size)
 	};
-	if (item->data == NULL) {
-		_CBOR_FREE(item);
-		return NULL;
-	}
+	_CBOR_DEPENDENT_NOTNULL(item, item->data);
+
 	return item;
 }
 
 cbor_item_t *cbor_new_indefinite_map()
 {
 	cbor_item_t *item = _CBOR_MALLOC(sizeof(cbor_item_t));
-	if (item == NULL)
-		return NULL;
+	_CBOR_NOTNULL(item);
+
 	*item = (cbor_item_t) {
 		.refcount = 1,
 		.type = CBOR_TYPE_MAP,