+++ src/cbor/strings.c
@@ -12,6 +12,7 @@
 cbor_item_t *cbor_new_definite_string()
 {
 	cbor_item_t *item = _CBOR_MALLOC(sizeof(cbor_item_t));
+	_CBOR_NOTNULL(item);
 	*item = (cbor_item_t) {
 		.refcount = 1,
 		.type = CBOR_TYPE_STRING,
@@ -23,12 +24,14 @@ cbor_item_t *cbor_new_definite_string()
 cbor_item_t *cbor_new_indefinite_string()
 {
 	cbor_item_t *item = _CBOR_MALLOC(sizeof(cbor_item_t));
+	_CBOR_NOTNULL(item);
 	*item = (cbor_item_t) {
 		.refcount = 1,
 		.type = CBOR_TYPE_STRING,
 		.metadata = {.string_metadata = {.type = _CBOR_METADATA_INDEFINITE, .length = 0}},
 		.data = _CBOR_MALLOC(sizeof(struct cbor_indefinite_string_data))
 	};
+	_CBOR_DEPENDENT_NOTNULL(item, item->data);
 	*((struct cbor_indefinite_string_data *) item->data) = (struct cbor_indefinite_string_data) {
 		.chunk_count = 0,
 		.chunk_capacity = 0,
@@ -40,8 +43,10 @@ cbor_item_t *cbor_new_indefinite_string()
 cbor_item_t *cbor_build_string(const char *val)
 {
 	cbor_item_t *item = cbor_new_definite_string();
+	_CBOR_NOTNULL(item);
 	size_t len = strlen(val);
 	unsigned char * handle = _CBOR_MALLOC(len);
+	_CBOR_DEPENDENT_NOTNULL(item, handle);
 	memcpy(handle, val, len);
 	cbor_string_set_handle(item, handle, len);
 	return item;
@@ -50,7 +55,9 @@ cbor_item_t *cbor_build_string(const char *val)
 cbor_item_t *cbor_build_stringn(const char *val, size_t length)
 {
 	cbor_item_t *item = cbor_new_definite_string();
+	_CBOR_NOTNULL(item);
 	unsigned char * handle = _CBOR_MALLOC(length);
+	_CBOR_DEPENDENT_NOTNULL(item, handle);
 	memcpy(handle, val, length);
 	cbor_string_set_handle(item, handle, length);
 	return item;